Function: Utility Function

$$U(R) = {R}^{\gamma}$$

Usage

func_gamma(reward, params, ...)

Arguments

reward

The feedback received by the agent from the environment at trial(t) following the execution of action(a)

params

Parameters used by the model’s internal functions, see params

...

It currently contains the following information; additional information may be added in future package versions.

idinfo:

subid

block

trial

exinfo: contains information whose column names are specified by the user.

Frame

RT

NetWorth

...

behave: includes the following:

action: the behavior performed by the human in the given trial.

latent: the object updated by the agent in the given trial.

simulation: the actual behavior performed by the agent.

cue and rsp: Cues and responses within latent learning rules, see behrule

Value

A NumericVector of length one representing the subjective value transformed from the objective reward via the utility function.

Body

func_gamma <- function(
    reward,
    params,
    ...
){

  list2env(list(...), envir = environment())

  # If you need extra information(...)
  # Column names may be lost(C++), indexes are recommended
  # e.g.
  # Trial  <- idinfo[3]
  # Frame  <- exinfo[1]
  # Action <- behave[1]

  gamma     <-  params[["gamma"]]

  # Stevens' Power Law
  utility <- sign(reward) * (abs(reward) ^ gamma)

  return(utility)
}