Dimension Reduction Methods (ABC)

Specifies the dimension reduction method for summary statistics in Approximate Bayesian Computation (ABC). High-dimensional summary statistics can lead to the "curse of dimensionality," where the algorithm struggles to find a solution. Reducing dimensions helps retain the "fingerprint" of the original data while removing noise, allowing the program to efficiently identify the underlying parameters.

Methods

NULL: No compression is applied. This is suitable for smaller datasets where the number of features (e.g., blocks * responses) is low (typically < 200). The ncomp parameter is ignored.

"PLS" (Partial Least Squares): A supervised method that compresses summary statistics into a lower-dimensional space defined by ncomp. It finds linear combinations of statistics that maximize covariance with the parameters, "guiding" the compression to prioritize information most relevant to parameter estimation.

"PCA" (Principal Component Analysis): An unsupervised method that compresses information into a lower-dimensional space defined by ncomp. It identifies orthogonal directions (principal components) that capture the maximum variance within the summary statistics themselves, preserving the data's most characteristic features without considering the parameters.

Related Parameters

ncomp [int] The number of components to retain after compression. By default, this is the number of blocks in the experiment. An excessive number of blocks or actions can create a high-dimensional summary space, making it hard for ABC to converge. Specifying an appropriate ncomp is crucial when using "PLS" or "PCA".

Example

 # supported reduction methods
 control = list(
   reduction = c(NULL, "PCA", "PLS"),
   ncomp = NULL
 )