Settings of Model

The settings argument is responsible for defining the model's name, the estimation method, and other configurations.

Class

settings [List]

Slot

name [Character]

The name of model.

mode [Character]

There are two modes: "fitting" and "simulating". In most cases, users do not need to explicitly specify the value of this slot, as the program will set it automatically.

Typically, the "fitting" mode is used when executing fit_p, while the "simulating" mode is used when executing rcv_d.

estimate [Character]

The package supports four estimation methods: MLE, MAP, ABC, and RNN. Generally, users no longer need to specify the estimation method in the settings object. This slot has been moved to an argument within the main functions, rcv_d and fit_p. For details, please refer to the documentation for estimate.

policy [Character]

The naming of this slot as policy is still under consideration.

Colloquially, policy = "on" means the agent selects an option based on its estimated probability and then updates the value of the chosen option.

Conversely, policy = "off" means the agent directly mimics human behavior, solely using its estimated probability and the human's choice to calculate the likelihood.

For details, please refer to the documentation for policy.

Example

 # model settings
 settings = list(
   name = "TD",
   mode = "fitting",
   estimate = "MLE",
   policy = "off"
 )